pyproject.toml
[tool.poetry]
name = "zarrita"
version = "0.1.0a1"
description = ""
authors = ["Norman Rzepka <code@normanrz.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
fsspec = "^2022.2.0"
numpy = "^1.24.2"
numcodecs = "^0.10.0"
cattrs = "^22.2.0"
attrs = "^22.2.0"
crc32c = "^2.3.post0"

[tool.poetry.group.dev.dependencies]
webknossos = "^0.12.3"
pytest = "^7.2.1"
black = "^23.1.0"
mypy = "^1.0.1"
isort = "^5.12.0"
pudb = "^2022.1.3"
zarr = "^2.14.2"
pytest-asyncio = "^0.21.0"

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
